fix(keybinds): replace ctrl with control on mac os
