+++ package.json
@@ -50,7 +50,6 @@
     "ed2curve": "^0.3.0",
     "emoji-mart-vue-fast": "^10.2.1",
     "events": "^3.3.0",
-    "faker": "^5.5.3",
     "filesize": "^8.0.7",
     "fp-ts": "^2.11.8",
     "fuse.js": "^6.5.3",
@@ -116,6 +115,7 @@
     "@types/howler": "^2.2.5",
     "@types/jest": "^27.4.1",
     "@types/lunr": "^2.3.4",
+    "@types/mousetrap": "^1.6.9",
     "@types/node": "^16.11.26",
     "@types/simple-peer": "^9.11.4",
     "@types/uuid": "^8.3.4",
@@ -139,6 +139,7 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-vue": "^8.5.0",
     "fake-indexeddb": "^3.1.7",
+    "faker": "5.5.3",
     "fp-ts": "^2.11.8",
     "husky": ">=6",
     "io-ts": "^2.2.16",
