+++ utilities/HotkeyList.ts
@@ -1,33 +1,3 @@
-export const keyboardRegex = /[a-z0-9\t\n /<>?;:"'`!@#$%^&*()[\]{}_+=|\\-]/gm
-
-export const specialKeys: { [key: string]: string } = {
-  control: 'ctrl',
-  shift: 'shift',
-  alt: 'alt',
-  option: 'option',
-  command: 'command',
-  meta: 'meta',
-  backspace: 'backspace',
-  tab: 'tab',
-  enter: 'enter',
-  return: 'return',
-  capslock: 'capslock',
-  esc: 'esc',
-  escape: 'escape',
-  space: 'space',
-  pageup: 'pageup',
-  pagedown: 'pagedown',
-  end: 'end',
-  home: 'home',
-  arrowup: 'up',
-  arrowright: 'right',
-  arrowleft: 'left',
-  arrowdown: 'down',
-  insert: 'ins',
-  delete: 'del',
-  plus: 'plus',
-}
-
 export const windowsShortcuts = [
   'ctrl+n',
   'ctrl+c',
@@ -154,7 +124,7 @@ export const macShortcuts = [
   'meta+alt+i',
   'meta+shift+delete',
   'meta+shift+m',
-  'ctrl+f2',
+  'control+f2',
   'meta+alt+arrowup',
   'meta+alt+shift+i',
   'f7',
@@ -173,7 +143,7 @@ export const macShortcuts = [
   'meta+alt=j',
   'meta+d',
   'meta+shift+d',
-  'meta+ctrl+f',
+  'meta+control+f',
   'meta++',
   'meta+-',
   'meta+0',
